second round of testing and tuning
This version uses a much simpler maneuvering algorithm, as
the approach initially taken doesn't work too well.

This approach is still sensitive to the amount of time it
takes to acquire LIDAR data. The device runs at about 3 Hz,
so at overscan=3 it takes about a second to run lidarScan().
This means the decision-making loop for stopping a turn will
often overshoot the heading it's looking for. overscan=1
seems to work best even though it may give ragged data.